+++ spec/component/augmented.spec.js
@@ -65,3 +65,28 @@ test('$html returns element html', function (t) {
   let vm = vuenit.component(component, options);
   t.is(vm.$html, vm.$el.outerHTML);
 });
+
+test('$create creates a new instance', async function (t) {
+  let {component, options} = t.context;
+  let vm1 = vuenit.component(component, options);
+  let vm2 = vm1.$create();
+
+  t.true(vm1 !== undefined);
+  t.not(vm1, vm2);
+  t.not(vm1.propsData, vm2.propsData);
+
+  t.is(vm1.dataA, vm2.dataA);
+  vm1.dataA = 'foo';
+  vm2.dataA = 'bah';
+  vm1.propsData.propA = 'C';
+  vm2.propsData.propA = 'D';
+  t.not(vm1.propsData.propA, vm2.propsData.propA);
+
+  await vm1.$nextTick();
+  await vm2.$nextTick();
+
+  t.not(vm1.dataA, vm2.dataA);
+
+  t.is(vm1.propA, 'C');
+  t.is(vm2.propA, 'D');
+});