+++ spec/helpers/setup.js
@@ -2,4 +2,4 @@ const browser = require('browser-env');
 const alias = require('module-alias');
 
 browser();
-alias.addAlias('vue', 'vue/dist/vue.min');
+alias.addAlias('vue', 'vue/dist/vue');